Renamed tests to be LoadDatabase
