+++ test/DatabaseIntegrationTest/Tests/LoadDatabase/LoadDatabaseTest.cpp
@@ -1,7 +1,7 @@
 #include "stdafx.h"
-#include "LoadEntityTestData.h"
+#include "LoadDatabaseTestData.h"
 
-#include "DatabaseIntegrationTest/Tests/LoadEntity/LoadUserScenariosBuilder.h"
+#include "DatabaseIntegrationTest/Tests/LoadDatabase/LoadUserScenariosBuilder.h"
 #include "DatabaseIntegrationTest/Tools/Core.h"
 #include "DatabaseIntegrationTest/Tools/SQLExecutor.h"
 
@@ -13,12 +13,12 @@ using namespace systelab::test_utility;
 
 namespace seed_cpp { namespace db_test {
 
-	class LoadEntityTest : public TestWithParam<LoadEntityTestData>
+	class LoadDatabaseTest : public TestWithParam<LoadDatabaseTestData>
 	{
 	public:
 		void SetUp()
 		{
-			m_databaseFilepath = "./LoadEntityTest.db";
+			m_databaseFilepath = "./LoadDatabaseTest.db";
 			m_sqlExecutor = std::make_unique<SQLExecutor>(m_databaseFilepath);
 			m_core = std::make_unique<Core>(m_databaseFilepath);
 
@@ -37,14 +37,14 @@ namespace seed_cpp { namespace db_test {
 		std::unique_ptr<Core> m_core;
 	};
 
-	TEST_P(LoadEntityTest, testLoadEntities)
+	TEST_P(LoadDatabaseTest, testLoadModelFromDatabase)
 	{
 		m_sqlExecutor->executeScripts(GetParam().m_sqlScripts);
 		m_core->initialize();
 		ASSERT_TRUE(EntityComparator()(GetParam().m_expectedModel, m_core->getModel()));
 	}
 
-	INSTANTIATE_TEST_CASE_P(LoadUsers,	LoadEntityTest, ValuesIn(LoadUserScenariosBuilder::build()));
+	INSTANTIATE_TEST_CASE_P(LoadUsers, LoadDatabaseTest, ValuesIn(LoadUserScenariosBuilder::build()));
 
 }}
