+++ test/DatabaseIntegrationTest/Tests/LoadDatabase/LoadUserScenariosBuilder.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Tests/LoadEntity/LoadEntityTestData.h"
+#include "Tests/LoadDatabase/LoadDatabaseTestData.h"
 
 
 namespace seed_cpp { namespace db_test {
@@ -8,7 +8,7 @@ namespace seed_cpp { namespace db_test {
 	class LoadUserScenariosBuilder
 	{
 	public:
-		static std::vector<LoadEntityTestData> build();
+		static std::vector<LoadDatabaseTestData> build();
 	};
 
 }}
